pyproject.toml
[tool.poetry]
name = "todo-or-else"
version = "0.1.0"
description = ""
authors = ["Jared Welch <jwelch92@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^3.9.2"
mypy = "^0.910"
pytest = "^6.2.5"
freezegun = "^1.1.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."flake8.extension"]
"DIE001" = "todo_or_else:flake8_entrypoint"

[tool.coverage.report]
fail_under = 80